+++ tests/test_official_doc.py
@@ -59,6 +59,13 @@ def test_build_html(app, status, warning):
     responses.add(responses.GET, re.compile(r"https://avatars.githubusercontent.com/.*"), body="")
     app.builder.build_all()
 
+    # Check if static files got copied correctly.
+    build_dir = Path(app.outdir) / "_static" / "sphinx-needs" / "libs" / "html"
+    files = [f for f in build_dir.glob("**/*") if f.is_file()]
+    assert build_dir / "sphinx_needs_collapse.js" in files
+    assert build_dir / "datatables_loader.js" in files
+    assert build_dir / "DataTables-1.10.16" / "js" / "jquery.dataTables.min.js" in files
+
 
 @responses.activate
 @with_app(buildername="singlehtml", srcdir="../docs")